orders add and get

+++ migrations/000002_create_orders_table.up.sql
@@ -1,6 +1,6 @@
 create table orders
 (
-    number     int not null constraint orders_pk primary key,
+    number     bigint not null constraint orders_pk primary key,
     user_id    integer constraint orders_users_id_fk references users,
     status     varchar(16) not null,
     created_at TIMESTAMP default CURRENT_TIMESTAMP,
